Feat: 상세 랭킹 조회 API 추가, 랭킹 조회 시 Redis 사용

+++ src/main/java/com/explorer/gabom/domain/quest/dto/response/QuestCreateResponse.java
@@ -17,8 +17,8 @@ public class QuestCreateResponse implements TargetIdentifiable {
 	private String description;
 	private QuestConditionType questConditionType;
 	private int acquireCondition;
-	private int rewardPoint;
-	private int rewardExp;
+	private Long rewardPoint;
+	private Long rewardExp;
 	private Long rewardTitleId;
 	private LocalDateTime createdAt;
