+++ src/main/java/com/explorer/gabom/domain/quest/entity/Quest.java
@@ -44,10 +44,10 @@ public class Quest extends BaseTimeEntity {
 	private int acquireCondition;
 
 	@Column(name = "reward_point", nullable = false)
-	private int rewardPoint;
+	private Long rewardPoint;
 
 	@Column(name = "reward_exp", nullable = false)
-	private int rewardExp;
+	private Long rewardExp;
 
 	@ManyToOne
 	@JoinColumn(name = "title_id")
@@ -60,7 +60,7 @@ public class Quest extends BaseTimeEntity {
 	private LocalDateTime deletedAt;
 
 	public Quest(String title, String description, QuestConditionType questConditionType,
-				 int acquireCondition, int rewardPoint, int rewardExp, Title rewardTitle) {
+				 int acquireCondition, Long rewardPoint, Long rewardExp, Title rewardTitle) {
 		this.title = title;
 		this.description = description;
 		this.questConditionType = questConditionType;
